# api/openapi.yaml
openapi: 3.1.0
info:
  title: Dhun API
  version: "1"
  description: |
    The only boundary between the Dhun server and its clients (AGENTS.md).
    Hand-written and reviewed like code; the server's
    TestSpecCoversEveryRoute fails if a route is missing here.

    Design and merge rules: docs/plans/006_api_and_sync.md.

    - Plain HTTP on the home network and over Tailscale; HTTPS optionally
      through `tailscale serve`. Request bodies must be `Content-Type: application/json`
      (else 415); responses are gzipped when the client sends
      `Accept-Encoding: gzip`.
    - Clients authenticate with `Authorization: Bearer <token>` from
      `POST /api/v1/login`. The `dhun_token` HttpOnly cookie that login also
      sets is accepted only by GET /stream, /art and /lyrics, for `<audio>`
      and `<img>` in the web client.
    - Browsers' cross-origin state-changing requests are refused (403).
    - Every signed-in response carries `Dhun-Version: 0.1.2`, the server's
      release ("dev" for a local build), for the apps to show. Requests
      without a valid token are not told.
    - Errors are `{"code": "...", "message": "..."}` with a matching status.
    - Every other GET outside `/api/` is the web client's files
      (docs/plans/030_web_client.md); an unknown `/api/` path is still a
      JSON 404.
servers:
  - url: http://{host}:8585
    variables:
      host:
        default: nas

security:
  - bearer: []

paths:
  /healthz:
    get:
      summary: Liveness and database check (no auth)
      security: []
      responses:
        "200": { description: "`ok`" }
        "503": { description: Database unavailable }

  /api/v1/login:
    post:
      summary: Sign in a device
      description: |
        Creates a device token. One token per app install; it stays valid
        until the device is removed (DELETE /devices/{id}, logout, or a
        password change).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, device]
              properties:
                username: { type: string }
                password: { type: string }
                device:
                  type: string
                  description: Shown in hand-off ("Continue from <device>")
                  example: Pixel
      responses:
        "200":
          description: Signed in; also sets the dhun_token cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  deviceId: { type: integer }
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Error" }
        "429":
          description: |
            Too many wrong passwords for this name (after five, each try
            waits twice as long, up to 15 minutes); see Retry-After.

  /api/v1/logout:
    post:
      summary: Sign out this device (deletes its token)
      responses:
        "204": { description: Signed out }

  /api/v1/me:
    get:
      summary: The signed-in user and device
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  deviceId: { type: integer }

  /api/v1/devices:
    get:
      summary: The caller's signed-in devices
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        name: { type: string }
                        createdAt: { type: string, format: date-time }
                        lastSeenAt: { type: string, format: date-time }
                        current: { type: boolean }

  /api/v1/devices/{id}:
    delete:
      summary: Sign out one of the caller's devices (a lost phone)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Signed out }
        "404": { $ref: "#/components/responses/Error" }

  /api/v1/library:
    get:
      summary: Songs and visible playlists changed since a cursor
      description: |
        Clients keep the whole catalogue and browse and search it locally.
        `since=0` (or omitted) returns everything; pass the returned
        `version` next time to get only changes. Missing songs come back
        with `missing: true`; removed playlists as `{id, deleted: true}`.
        Playlists are the shared ones plus the caller's own.
      parameters:
        - name: since
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  version: { type: integer }
                  songs:
                    type: array
                    items: { $ref: "#/components/schemas/Song" }
                  playlists:
                    type: array
                    items: { $ref: "#/components/schemas/Playlist" }

  /api/v1/stream/{id}:
    get:
      summary: The original audio file, with HTTP Range support
      description: Never transcoded. Downloads use this endpoint too.
      security: [{ bearer: [] }, { cookie: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Whole file }
        "206": { description: Requested range }
        "404": { $ref: "#/components/responses/Error" }

  /api/v1/art/{id}:
    get:
      summary: A song's cover art
      description: Embedded art first, else the folder's cover file. Supports ETag.
      security: [{ bearer: [] }, { cookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - name: size
          in: query
          description: |
            Longest side in pixels (JPEG), rounded up to 128, 256, 512 or
            1024; 0, absent or larger for the original. Images over 40
            megapixels are always served as they are.
          schema: { type: integer, minimum: 0 }
      responses:
        "200": { description: Image bytes }
        "304": { description: Not modified }
        "404": { $ref: "#/components/responses/Error" }

  /api/v1/thumbs:
    post:
      summary: Small covers by art key
      description: |
        128 px covers (JPEG, longest side), base64, for the art keys a client
        does not have yet: it keeps one for every cover in the library, so
        list rows never wait for art (docs/plans/019_networking_and_caching.md).
        A key with nothing to show maps to "", so the client stops asking;
        one that failed for another reason is left out, to be asked again.
        At most 50 keys a request: the first time, each may mean decoding a
        large cover.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [keys]
              properties:
                keys:
                  { type: array, items: { type: string }, minItems: 1, maxItems: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  thumbs:
                    type: object
                    additionalProperties: { type: string }
        "400": { $ref: "#/components/responses/Error" }

  /api/v1/lyrics/{id}:
    get:
      summary: A song's lyrics, as stored
      description: |
        The sibling .lrc file if there is one, else embedded lyrics.
        Clients parse LRC timestamps themselves.
      security: [{ bearer: [] }, { cookie: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  source: { type: string, enum: [lrc, embedded] }
                  synced: { type: boolean }
                  text: { type: string }
        "404": { $ref: "#/components/responses/Error" }

  /api/v1/search/lyrics:
    get:
      summary: Songs whose lyrics hold a search, with the line that does
      description: |
        A line matches when it holds every word of `q`, folded as
        api/search-fold.tsv shows: by the whole word, its start, inside it,
        or the start of its sound key. Lines holding the words as typed come
        first; at most 50 songs, one line each. Everything else is searched
        on the device (docs/plans/029_search.md).
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  hits:
                    type: array
                    items:
                      type: object
                      properties:
                        song: { type: integer, format: int64 }
                        line:
                          {
                            type: string,
                            description: The line as written,
                            without LRC times,
                          }

  /api/v1/sync:
    get:
      summary: The caller's queues, special lists, resume points, settings and now-playing changed since a cursor
      parameters:
        - name: since
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncState" }
    post:
      summary: Push operations, then pull changes, in one round trip
      description: |
        Operations are applied in order, each atomically. Each has a
        client-generated `id`; a resent id is reported as `duplicate` and
        not applied again. A `rejected` operation can never succeed and
        must be dropped from the client's outbox. Playlist operations
        rewrite the .m3u8 file and are visible through /library.

        At most 500 operations are applied per request: the rest get no
        result, stay in the outbox and go with the next sync. An operation
        may name at most 20,000 songs. An `at` in the future counts as the
        server's now.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ops]
              properties:
                since: { type: integer, description: The client's sync cursor }
                ops:
                  type: array
                  items: { $ref: "#/components/schemas/Op" }
      responses:
        "200":
          description: Results, plus everything changed since `since`
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SyncState" }
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string }
                            status:
                              { type: string, enum: [applied, duplicate, rejected] }
                            error: { type: string }

  /api/v1/now-playing:
    get:
      summary: The caller's latest playback, for resume hand-off
      description: |
        An app coming to the foreground offers "Continue from <deviceName>"
        when this is newer than its own playback and from another device.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  nowPlaying:
                    oneOf:
                      - { $ref: "#/components/schemas/NowPlaying" }
                      - { type: "null" }

  /api/v1/plays:
    get:
      summary: The caller's play count and last play per song
      description: |
        Counts only listens of at least half the song; every listen is
        still logged (docs/plans/008_listening_history.md).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  plays:
                    type: array
                    items:
                      type: object
                      properties:
                        song: { type: integer }
                        count: { type: integer }
                        lastPlayedAt: { type: string, format: date-time }

  /api/v1/admin/users:
    get:
      summary: Every user, with their device count (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/User"
                        - type: object
                          properties:
                            createdAt: { type: string, format: date-time }
                            devices: { type: integer }
                            lastSeenAt:
                              type: string
                              format: date-time
                              description: Omitted when no device is signed in
        "403": { $ref: "#/components/responses/Error" }
    post:
      summary: Add a family member (admin)
      description: |
        Creates a member. The only admin is the one created from
        DHUN_ADMIN_USER and DHUN_ADMIN_PASSWORD on first start. The name is
        also the user's `Playlists/<name>/` folder, so it must be usable as a
        folder name; names are case-insensitive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, password]
              properties:
                name: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }

  /api/v1/admin/users/{id}/password:
    put:
      summary: Reset a user's password (admin); signs out all their devices
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        "204": { description: Changed }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }

  /api/v1/admin/rescan:
    post:
      summary: Start a library scan now (admin)
      responses:
        "202": { description: Scan queued }
        "403": { $ref: "#/components/responses/Error" }

components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    cookie: { type: apiKey, in: cookie, name: dhun_token }

  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }

  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string, example: not_found }
              message: { type: string }

  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        admin: { type: boolean }

    Song:
      type: object
      description: Optional text fields are omitted when empty.
      required:
        [id, path, title, artists, genres, durationMs, size, hasArt, hasLyrics, addedAt]
      properties:
        id:
          {
            type: integer,
            description: Stable; never reused,
            survives moves and upgrades,
          }
        path:
          {
            type: string,
            description: Relative to the Music folder; for folder browsing,
          }
        title: { type: string }
        artist: { type: string, description: Raw tag }
        artists:
          { type: array, items: { type: string }, description: Split on, ; & feat. }
        album: { type: string }
        albumArtist: { type: string }
        composer: { type: string }
        genre: { type: string }
        genres: { type: array, items: { type: string } }
        year: { type: integer }
        track: { type: integer }
        disc: { type: integer }
        durationMs: { type: integer }
        format:
          type: string
          example: mpeg
          description: >
            The container, read from the file's content, not its name: mpeg,
            mp4, flac, ogg, wav, aiff, webm or matroska. A WebM file saved
            under an .opus name is webm. Empty when the file could not be read.
        codec:
          type: string
          example: aac
          description: The codec inside a container (aac, alac, opus, vorbis, pcm); empty for mpeg and flac
        bitrate: { type: integer, description: kbit/s }
        sampleRate: { type: integer }
        bitDepth: { type: integer }
        size: { type: integer, description: Bytes; for download quotas }
        hasArt: { type: boolean }
        art:
          type: string
          description: >
            Which cover the song shows: songs showing the same cover share
            the key, and a new cover gets a new one, so a client keeps covers
            by key without asking again (docs/plans/019_networking_and_caching.md).
            Fetch one through /art/{id} of any song with that key. Omitted
            without art.
        hasLyrics: { type: boolean }
        addedAt: { type: string, format: date-time }
        missing:
          { type: boolean, description: File gone; keep the song's data, hide it }

    Playlist:
      type: object
      properties:
        id: { type: integer }
        deleted: { type: boolean }
        path: { type: string }
        name: { type: string }
        shared:
          {
            type: boolean,
            description: Top-level Playlists/ file; only the admin edits it,
          }
        ref:
          {
            type: string,
            description: "The ref from the playlist.create that made it, so the creating app can match a playlist it made offline",
          }
        songs:
          type: array
          items: { type: integer }
          description: In order; 0 is an entry that matches no song (kept, shown as unavailable)

    Queue:
      type: object
      properties:
        id: { type: string, description: Client-generated }
        deleted: { type: boolean }
        name: { type: string }
        songs:
          {
            type: array,
            items: { type: integer },
            description: Play order; no duplicates,
          }
        currentSong: { type: integer }
        positionMs: { type: integer }
        shuffle: { type: boolean }
        repeat: { type: string, enum: ["off", queue, song] }
        usedAt: { type: string, format: date-time }

    NowPlaying:
      type: object
      properties:
        deviceId: { type: integer }
        deviceName: { type: string }
        queue: { type: string }
        song: { type: integer }
        positionMs: { type: integer }
        playing: { type: boolean }
        at: { type: string, format: date-time }

    ListItem:
      type: object
      properties:
        song: { type: integer }
        deleted: { type: boolean }
        at: { type: string, format: date-time, description: When it was added }

    SyncState:
      type: object
      properties:
        version: { type: integer, description: Next cursor }
        queues: { type: array, items: { $ref: "#/components/schemas/Queue" } }
        favorites:
          description: Newest first by `at` (docs/plans/010_special_playlists.md)
          type: array
          items: { $ref: "#/components/schemas/ListItem" }
        listenLater:
          description: Same as favorites; the server removes an item a listen reached 90% into
          type: array
          items: { $ref: "#/components/schemas/ListItem" }
        resume:
          description: Where long files were left (docs/plans/009_resume_long_files.md)
          type: array
          items:
            type: object
            properties:
              song: { type: integer }
              deleted: { type: boolean, description: Finished; forget it }
              positionMs: { type: integer }
              at: { type: string, format: date-time }
        settings:
          description: |
            App settings changed since the cursor, as name → JSON value. A
            missing setting or null means the app's default.
            `longFiles.minMinutes` (default 15) and `longFiles.resume`
            ("auto", "ask" or "off"; default "auto"), read by the apps;
            `listenLater.autoRemove` (default true) and
            `listenLater.finishedPercent` (default 90), read by the server;
            `speed.<song id>` (`{"speed": 1.5, "semitones": 0}`), a song's
            own play speed and pitch, read by the apps.
          type: object
          additionalProperties: true
        nowPlaying:
          oneOf:
            - { $ref: "#/components/schemas/NowPlaying" }
            - { type: "null" }

    Op:
      type: object
      description: |
        One change recorded by a client. `after` positions relative to a
        neighbouring song: absent = end, 0 = start, unknown song = end.

        | type | fields | merge rule |
        | --- | --- | --- |
        | queue.create | queue, name, songs, song?, positionMs? | name made unique; past 20 queues the least recently used is removed |
        | queue.rename | queue, name | name made unique |
        | queue.delete | queue | delete wins over later edits |
        | queue.insert | queue, songs, after? | songs already queued are moved, never duplicated |
        | queue.remove | queue, songs | |
        | queue.move | queue, song, after? | |
        | queue.replace | queue, songs | whole order (sort, shuffle); last writer wins |
        | queue.set_current | queue, song, positionMs | later `at` wins |
        | queue.set_mode | queue, shuffle?, repeat? | |
        | favorite.set / favorite.unset | song | later `at` wins |
        | listen_later.add / listen_later.remove | song | later `at` wins |
        | resume.set / resume.unset | song, positionMs | later `at` wins; where a long file was left |
        | setting.set | name, value | later `at` wins; name 1–64 of [A-Za-z0-9._-], value any JSON ≤ 4 KB, ≤ 1000 per user not counting nulls |
        | play | song, ms, endedAt?, fromMs?, toMs?, end?, source?, queue?, shuffle?, utcOffset? | append only; one per listen, skips included; `at` is when it started; a `toMs` past 90% removes the song from Listen Later |
        | playback.state | queue, song, positionMs, playing | later `at` wins; feeds /now-playing |
        | playlist.create | ref, name, songs | file Playlists/<user>/<name>.m3u8; "Favorites" and "Listen Later" are reserved |
        | playlist.rename | playlist, name | renames the file |
        | playlist.delete | playlist | file removed after a copy is kept in the data folder |
        | playlist.insert | playlist, songs, after?, afterOccurrence? | duplicates allowed |
        | playlist.remove | playlist, song, occurrence? | |
        | playlist.move | playlist, song, occurrence?, after?, afterOccurrence? | |
        | playlist.replace | playlist, songs | |

        Playlists: your own, or shared ones if you are the admin.
      required: [id, type]
      properties:
        id: { type: string, description: Client-generated, unique per user }
        at:
          {
            type: string,
            format: date-time,
            description: When the change was made; defaults to receipt time,
          }
        type: { type: string }
        queue: { type: string }
        playlist:
          {
            type: string,
            description: 'Server id ("123") or "ref:<ref>" for one created offline',
          }
        ref: { type: string }
        name: { type: string }
        songs: { type: array, items: { type: integer } }
        song: { type: integer }
        after: { type: integer }
        occurrence: { type: integer, description: Which copy of song (0-based) }
        afterOccurrence: { type: integer }
        positionMs: { type: integer }
        playing: { type: boolean }
        shuffle: { type: boolean }
        repeat: { type: string, enum: ["off", queue, song] }
        ms: { type: integer, description: "play: time heard, pauses excluded" }
        endedAt: { type: string, format: date-time }
        fromMs: { type: integer, description: Where in the song the listen started }
        toMs: { type: integer, description: Where it ended }
        end:
          {
            type: string,
            description: "finished, skipped, previous, switched, stopped or interrupted; other values are kept as given",
          }
        source:
          {
            type: string,
            description: "What the queue was started from: album:<id>, artist:<name>, playlist:<id>, folder:<path>, favorites, search",
          }
        utcOffset: { type: integer, description: "The device's UTC offset, in minutes" }
        value: { description: "setting.set: any JSON" }